services/llm-analysis/pyproject.toml
[project]
name = "llm-analysis"
version = "0.1.0"
description = "LLM-powered email analysis service for mail-manager"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "structlog>=24.4.0",
    "asyncpg>=0.30.0",
    "httpx>=0.28.0",
    "redis[hiredis]>=5.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_analysis"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "docker: tests that require Docker daemon (build and run containers)",
]
